+++ src/ZEO/tests/testConnection.py
@@ -27,8 +27,11 @@ if os.environ.get('USE_ZOPE_TESTING_DOCTEST'):
 else:
     import doctest
 import unittest
-import ZEO.tests.forker
 import ZODB.tests.util
+
+import ZEO
+
+from . import forker
 
 class FileStorageConfig:
     def getConfig(self, path, create, read_only):
@@ -79,12 +82,6 @@ class MappingStorageConnectionTests(
     ):
     """Mapping storage connection tests."""
 
-class SSLConnectionTests(
-    MappingStorageConfig,
-    ConnectionTests.SSLConnectionTests,
-    ):
-    pass
-
 # The ReconnectionTests can't work with MappingStorage because it's only an
 # in-memory storage and has no persistent state.
 
@@ -107,8 +104,9 @@ test_classes = [FileStorageConnectionTests,
                 FileStorageTimeoutTests,
                 MappingStorageConnectionTests,
                 MappingStorageTimeoutTests,
-                SSLConnectionTests,
                 ]
+if not forker.ZEO4_SERVER:
+    test_classes.append(SSLConnectionTests)
 
 def invalidations_while_connecting():
     r"""
@@ -129,7 +127,7 @@ This tests tries to provoke this bug by:
 - opening a client to the server that writes some objects, filling
   it's cache at the same time,
 
-    >>> import ZODB.tests.MinPO, transaction
+    >>> import ZEO, ZODB.tests.MinPO, transaction
     >>> db = ZEO.DB(addr, client='x')
     >>> conn = db.open()
     >>> nobs = 1000
@@ -205,9 +203,9 @@ This tests tries to provoke this bug by:
     ...                          record = handler.records.pop(0)
     ...                          print(record.name, record.levelname, end=' ')
     ...                          print(handler.format(record))
-    ...        if bad:
-    ...           with open('server-%s.log' % addr[1]) as f:
-    ...               print(f.read())
+    ...        #if bad:
+    ...        #   with open('server.log') as f:
+    ...        #       print(f.read())
     ...        #else:
     ...        #   logging.getLogger('ZEO').debug('GOOD %s' % c)
     ...        db.close()
@@ -236,7 +234,7 @@ def test_suite():
         sub = unittest.makeSuite(klass, 'check')
         suite.addTest(sub)
     suite.addTest(doctest.DocTestSuite(
-        setUp=ZEO.tests.forker.setUp, tearDown=setupstack.tearDown,
+        setUp=forker.setUp, tearDown=setupstack.tearDown,
         ))
     suite.layer = ZODB.tests.util.MininalTestLayer('ZEO Connection Tests')
     return suite