+++ src/ZEO/tests/test_client_side_conflict_resolution.py
@@ -9,6 +9,7 @@ from ZODB.broken import find_global
 
 import ZEO
 
+from . import forker
 from .utils import StorageServer
 
 class Var(object):
@@ -16,10 +17,10 @@ class Var(object):
         self.value = other
         return True
 
+@unittest.skipIf(forker.ZEO4_SERVER, "ZEO4 servers don't support SSL")
 class ClientSideConflictResolutionTests(zope.testing.setupstack.TestCase):
 
     def test_server_side(self):
-
         # First, verify default conflict resolution.
         server = StorageServer(self, DemoStorage())
         zs = server.zs