+++ src/ZEO/tests/testssl.py
@@ -10,6 +10,7 @@ from ..Exceptions import ClientDisconnected
 from .. import runzeo
 
 from .testConfig import ZEOConfigTestBase
+from . import forker
 
 here = os.path.dirname(__file__)
 server_cert = os.path.join(here, 'server.pem')
@@ -19,6 +20,7 @@ serverpw_key  = os.path.join(here, 'serverpw_key.pem')
 client_cert = os.path.join(here, 'client.pem')
 client_key  = os.path.join(here, 'client_key.pem')
 
+@unittest.skipIf(forker.ZEO4_SERVER, "ZEO4 servers don't support SSL")
 class SSLConfigTest(ZEOConfigTestBase):
 
     def test_ssl_basic(self):
@@ -114,6 +116,7 @@ class SSLConfigTest(ZEOConfigTestBase):
             )
         stop()
 
+@unittest.skipIf(forker.ZEO4_SERVER, "ZEO4 servers don't support SSL")
 @mock.patch(('asyncio' if PY3 else 'trollius') + '.async')
 @mock.patch(('asyncio' if PY3 else 'trollius') + '.set_event_loop')
 @mock.patch(('asyncio' if PY3 else 'trollius') + '.new_event_loop')